F: add std feature
