+++ src/types.rs
@@ -1,9 +1,11 @@
+#[cfg(feature = "tokio")]
 #[derive(Debug)]
 pub struct Message<T, R> {
     inner: T,
     result_tx: Option<tokio::sync::oneshot::Sender<R>>,
 }
 
+#[cfg(feature = "tokio")]
 impl<T, R> Message<T, R>
 where
     T: Sized + Send + Clone,
@@ -18,6 +20,32 @@ where
     }
 
     pub fn result_tx(self) -> Option<tokio::sync::oneshot::Sender<R>> {
+        self.result_tx
+    }
+}
+
+#[cfg(feature = "std")]
+#[derive(Debug)]
+pub struct Message<T, R> {
+    inner: T,
+    result_tx: Option<oneshot::Sender<R>>,
+}
+
+#[cfg(feature = "std")]
+impl<T, R> Message<T, R>
+where
+    T: Sized + Send + Clone,
+    R: Sized + Send,
+{
+    pub fn new(inner: T, result_tx: Option<oneshot::Sender<R>>) -> Self {
+        Self { inner, result_tx }
+    }
+
+    pub fn inner(&self) -> T {
+        self.inner.clone()
+    }
+
+    pub fn result_tx(self) -> Option<oneshot::Sender<R>> {
         self.result_tx
     }
 }
